requirements.txt
Django>=2.0
django-pgconninfo>=0.9.1
